Refactor use getDecomposedIdentity
