+++ src/fontra/client/core/var-glyph.js
@@ -1,3 +1,4 @@
+import { getDecomposedIdentity } from "./transform.js";
 import { VarPackedPath } from "./var-path.js";
 
 export class VariableGlyph {
@@ -63,22 +64,10 @@ export class StaticGlyph {
   }
 }
 
-const identityTransformation = {
-  translateX: 0,
-  translateY: 0,
-  rotation: 0,
-  scaleX: 1,
-  scaleY: 1,
-  skewX: 0,
-  skewY: 0,
-  tCenterX: 0,
-  tCenterY: 0,
-};
-
 export function copyComponent(component) {
   return {
     name: component.name,
-    transformation: { ...identityTransformation, ...component.transformation },
+    transformation: { ...getDecomposedIdentity(), ...component.transformation },
     location: { ...component.location },
   };
 }
