+++ test-js/test-var-glyph.js
@@ -1,5 +1,6 @@
 import { expect } from "chai";
 
+import { getDecomposedIdentity } from "../src/fontra/client/core/transform.js";
 import { enumerate } from "../src/fontra/client/core/utils.js";
 import { StaticGlyph, VariableGlyph } from "../src/fontra/client/core/var-glyph.js";
 
@@ -27,17 +28,7 @@ function makeTestGlyphObject() {
             {
               name: "test",
               location: { a: 0.5 },
-              transformation: {
-                translateX: 0,
-                translateY: 0,
-                rotation: 0,
-                scaleX: 1,
-                scaleY: 1,
-                skewX: 0,
-                skewY: 0,
-                tCenterX: 0,
-                tCenterY: 0,
-              },
+              transformation: getDecomposedIdentity(),
             },
           ],
         },
@@ -68,17 +59,7 @@ describe("var-glyph Tests", () => {
       components: [
         {
           name: "test",
-          transformation: {
-            rotation: 0,
-            scaleX: 1,
-            scaleY: 1,
-            skewX: 0,
-            skewY: 0,
-            tCenterX: 0,
-            tCenterY: 0,
-            translateX: 0,
-            translateY: 0,
-          },
+          transformation: getDecomposedIdentity(),
           location: {},
         },
       ],
